requirements.txt
astroid==2.0.4
autopep8==1.4.2
dj-database-url==0.5.0
Django==2.0.2
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.4
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
python-decouple==3.1
pytz==2018.3
six==1.11.0
wrapt==1.10.11
